revert: feat(http,http2): apply http.endpoint and queryStringObfuscation to client spans (#8407)
This reverts commit 55ed50f87ed6f8af12a6598df76ccf599fe018ad.

stripping the outbound query string unconditionally and instead redacted it
with `queryStringObfuscation`. For high-traffic clients with variable query
strings (UUIDs, ISO timestamps, free-text search), retaining the query string
— even fully redacted — reintroduced a large source of `http.url` cardinality
that the previous strip-all behaviour did not have. A customer hit exactly this
explosion.

Reverting restores the pre-#8407 client behaviour (query string stripped from
`http.url`, no `http.endpoint` on client spans) while we decide whether the
parity feature is worth the cardinality cost and, if so, how to bound it
(e.g. type-token redaction rather than verbatim values).

The server side is untouched by this revert.

Refs: https://github.com/DataDog/dd-trace-js/issues/2022
