+++ index.d.ts
@@ -2017,8 +2017,7 @@ declare namespace tracer {
       middleware?: boolean;
 
       /**
-       * Whether (or how) to obfuscate querystring values in `http.url` on both
-       * inbound (server) and outbound (client) HTTP spans.
+       * Whether (or how) to obfuscate querystring values in `http.url`.
        *
        * - `true`: obfuscate all values
        * - `false`: disable obfuscation
@@ -2097,8 +2096,7 @@ declare namespace tracer {
        */
       validateStatus?: (code: number) => boolean;
       /**
-       * Whether (or how) to obfuscate querystring values in `http.url` on both
-       * inbound (server) and outbound (client) HTTP spans.
+       * Whether (or how) to obfuscate querystring values in `http.url`.
        *
        * - `true`: obfuscate all values
        * - `false`: disable obfuscation
