+++ index.d.v5.ts
@@ -2143,8 +2143,7 @@ declare namespace tracer {
       middleware?: boolean;
 
       /**
-       * Whether (or how) to obfuscate querystring values in `http.url` on both
-       * inbound (server) and outbound (client) HTTP spans.
+       * Whether (or how) to obfuscate querystring values in `http.url`.
        *
        * - `true`: obfuscate all values
        * - `false`: disable obfuscation
@@ -2223,8 +2222,7 @@ declare namespace tracer {
        */
       validateStatus?: (code: number) => boolean;
       /**
-       * Whether (or how) to obfuscate querystring values in `http.url` on both
-       * inbound (server) and outbound (client) HTTP spans.
+       * Whether (or how) to obfuscate querystring values in `http.url`.
        *
        * - `true`: obfuscate all values
        * - `false`: disable obfuscation
